requirements.txt
asn1crypto==0.24.0
autopep8==1.4.4
backports.csv==1.0.7
certifi==2019.3.9
cryptography==2.2.2
defusedxml==0.6.0
diff-match-patch==20121119
Django==2.0.10
django-cors-headers==2.3.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-reversion==3.0.3
django-suit==0.2.26
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.7
freeze==1.0.10
future==0.15.2
httplib2==0.9.2
idna==2.7
jdcal==1.4
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.4
psutil==5.6.3
pycodestyle==2.5.0
pycparser==2.18
pyflakes==2.1.1
PyJWT==1.6.4
PyMySQL==0.9.2
pytz==2018.5
PyYAML==3.13
six==1.11.0
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
wincertstore==0.2
xlrd==1.1.0
xlwt==1.3.0
pymongo==3.8.0
